[rl] PR review: rename grpo.py, (step, group_idx) env addressing
- Rename grpo_sync.py -> grpo.py (Danielle): "sync" wasn't legible from
  the filename; async loop will live in a separate module when it lands.
- Rename _collect_rollout -> _collect_rollouts (Tianyu): plural matches
  the returned list and follows SB3's collect_rollouts convention.
- Drop implicit grouping in _collect_rollouts: each Completion already
  carries prompt_idx, so trajectories can be built in a single pass with
  no slicing or group_size arithmetic.
- Switch env addressing from build(idx=i) to build(step=, group_idx=).
  The flat-int idx scheme made the script compute step * num_groups + i,
  leaking the env's addressing assumption. The new contract hands the env
  scheduling context only and lets the env decide what problem belongs
  there.
- README: note the loop is single-turn only.
