+++ .github/workflows/integration_test_4gpu_rl.yaml
@@ -80,7 +80,7 @@ jobs:
         uv pip install torchdata==0.12.0.dev20260327 --extra-index-url https://download.pytorch.org/whl/nightly/cpu
 
         # 4. Install torchtitan in editable mode because integration test spawns
-        # python grpo_sync.py as a subprocess
+        # python grpo.py as a subprocess
         uv pip install -e .
 
         # 5. Download HF model checkpoint for tests
